Refactors ArtObjectList into functional component and moves filtering to Gallery
