+++ src/components/Gallery.js
@@ -1,13 +1,39 @@
-import React from 'react';
+import React, { useState } from 'react';
 import art from '../art.json';
 import ArtObjectList from './ArtObjectList';
 
-function App() {
+const style = {
+  display: 'flex',
+  flexDirection: 'row',
+  justifyContent: 'space-between',
+  alignItems: 'center',
+};
+
+const Gallery = () => {
+  const [featuredOnly, setFeaturedOnly] = useState(false);
+
+  const filteredArtObjects = !featuredOnly
+    ? art
+    : art.filter((artObject) => artObject.featured);
+  const countMessage = featuredOnly
+    ? `${filteredArtObjects.length} Featured Objects`
+    : `${filteredArtObjects.length} Objects`;
   return (
     <main>
-      <ArtObjectList artObjects={art} />
+      <header style={style}>
+        <div>
+          <p>{countMessage}</p>
+        </div>
+
+        <div>
+          <button onClick={() => setFeaturedOnly((prev) => !prev)}>
+            {featuredOnly ? 'Show All' : 'Show Featured Only'}
+          </button>
+        </div>
+      </header>
+      <ArtObjectList artObjects={filteredArtObjects} />
     </main>
   );
-}
+};
 
-export default App;
+export default Gallery;
